Refactor Main and Launcher class
